Remove Flyway configuration and tests
Delete custom Flyway bean and its tests; remove Flyway dependency and config property; update CI/test settings. Removed MigrationConfiguration and MigrationConfigurationTests, removed flyway-database-postgresql from pom.xml, removed spring.flyway.enabled from application.properties, removed SPRING_FLYWAY_ENABLED env in GitHub Actions, and removed the test property that disabled Flyway in IntervuApplicationTests.

+++ src/main/java/com/intervu/interview/InterviewRepository.java
@@ -190,13 +190,13 @@ public class InterviewRepository {
 	}
 
 	public void insertEvent(UUID sessionId, String eventType, String payload) {
-		Integer nextVersion = jdbcTemplate.queryForObject(
+		Long nextVersion = jdbcTemplate.queryForObject(
 			"""
 				SELECT COALESCE(MAX(event_version), 0) + 1
 				FROM session_events
 				WHERE session_id = ?
 				""",
-			Integer.class,
+			Long.class,
 			sessionId
 		);
 		jdbcTemplate.update(
@@ -206,7 +206,7 @@ public class InterviewRepository {
 				""",
 			UUID.randomUUID(),
 			sessionId,
-			nextVersion == null ? 1 : nextVersion,
+			nextVersion == null ? 1L : nextVersion,
 			eventType,
 			payload
 		);
